+++ src/player/mod.rs
@@ -10,8 +10,8 @@ pub use alphabetasearchplayer::*;
 
 use crate::game;
 
-pub const MIN_SCORE: isize = isize::MIN + 1;
-pub const MAX_SCORE: isize = isize::MAX;
+pub const MIN_SCORE: i64 = i64::MIN + 1;
+pub const MAX_SCORE: i64 = i64::MAX;
 
 pub trait Player {
     fn get_move<'a>(&mut self, board: &mut game::Board, possible_moves: &'a Vec<game::Move>) -> Option<&'a game::Move>;
